Added some player, enemy and npc animations and npc bodies
Player death animation, melee enemy attack, npc  bodies and animation idle(breathing) and talking.
NPC dialogues need to be fixed.

In my last commit i managed the player attack with mouse click, the animation and the delay for the next attack, now the code has been changed a bit and the attack doesn't work as well as before