Fixed bug : Power up and cheats disturbing each other
Before when a activate a power up and then i activate the cheats, each of them increased the stats and it was too much.

I should have fixed that.

To be safe, check if taking the fire fist power up, activate and deactivate the cheats immediately after, the damage stats works properly.